Better sizing of Gantt chart parts.  Refs #8529
 * Allow setting of width of left table (name, duration, etc.)

Parameter name is 'lwidth'.  If omitted, the old jsGantt default, 585,
prevails. That comes from 15 for the +/- control, 220 for name, and 70
for each of the default columns (resource, %complete, etc.).

It used to be that name grew as other columns were removed..  Now the
table shrinks.

 * Make Gantt chart (on right) fill page.

The old structure was weird.  It had

 * div (for the whole Gantt)
  * table
   * tbody
    * tr
     * td (to hold the left side)
      * div (to style the left side)
       * the left table with name, duration, etc.
     * td (to hold the right side)
      * div (to style the left side)
       * the Gantt chart

The table made it unnessarily hard to format the Gantt to fill the
available space.

http://stackoverflow.com/questions/1017880/expand-div-to-max-width-when-floatleft-is-set shows how to do it with just <div>s.  I simplified the structure to:

 * div (for the whole Gantt)
  * div (to style the left side)
   * the left table with name, duration, etc.
  * div (to style the left side)
   * the Gantt chart

And updated the style sheet per the recommendations in the link above.


git-svn-id: http://trac-hacks.org/svn/tracjsganttplugin/0.11@10041 7322e99d-02ea-0310-aa39-e9a107903beb