Add `root=` option to jsGantt.
This allows you to display a ticket and all of its descendants.

I'm not entirely happy with the function names; _children() actually
returns the whole tree.  But this works.

Fix crash on missing dependency which turned up when I queried by `root=`.

We may want to bring in these tickets instead of skipping them.  Or
maybe that should be an option.


git-svn-id: http://trac-hacks.org/svn/tracjsganttplugin/0.11@10042 7322e99d-02ea-0310-aa39-e9a107903beb