Schedule tasks by hour, not day.  Refs #9042, #8982.
This is an first step toward handling dependencies (so, a four-hour task
and an 12-hour task can take just two 8-hour days) but it also addresses
a display but that tasks that took a multiple of 8 hours showed up a day
too long in the Gantt.

Giving Trac milestones 0 estimates simplifies other code.

git-svn-id: http://trac-hacks.org/svn/tracjsganttplugin/0.11@10690 7322e99d-02ea-0310-aa39-e9a107903beb