Pass scheduler tickets by ID, not unordered list.  Refs #9566.
Any scheduler will need a fast lookup of tickets by ID so this takes the
work of building that lookup off the schedulers, reducing duplicated
code.

It also copies the ticket attributes so that the scheduler can change
them without concern that it will break things.  This enforces the pre-
and post-conditions for `computeSchedule()` that only 'calc_start' and
'calc_finish' are set/changed.

It would seem that I could done

{{{
        for t in tickets:
            ticketsByID[t['id']] = copy.deepcopy(t)
}}}

but that raised errors.

git-svn-id: http://trac-hacks.org/svn/tracjsganttplugin/0.11@11143 7322e99d-02ea-0310-aa39-e9a107903beb