Tasks inherit their ancestors' dependencies.  Refs #9566.
If A has children B and C, and D has children E and F, if A is required
for D then B and C have to be completed before D.

This copies predecessor and successor dependencies down the tree before
scheduling so those dependencies can be honored by the scheduler.

It uses functions as objects and takes advantage of the fact that the
accessor functions don't return values, they return references to the
original lists so those lists can be updated without more complicated
access wrappers to append, remove, etc.

Also fix a bug in creating milestone pseudotickets where [] and None
were not used consistently with the rest of the module.

git-svn-id: http://trac-hacks.org/svn/tracjsganttplugin/0.11@11144 7322e99d-02ea-0310-aa39-e9a107903beb