Try to propagate dependencies to all children.  Refs #9163.
This actually seems to get back to the original implementation.  I broke
it trying to optimized.

The logic now and originally is:

{{{
for each child
   if the child doesn't depend on any cousins
      copy dependencies down
   recurse to process grandchildren
}}}

The optimized, broken, intermediate version did:

{{{
if there are dependencies
   for each child
      copy dependencies down
      recurse to process grandchildren
}}}

If any level of the tree had no dependencies, propagation was
interrupted. If a "root" ticket had no dependencies, all of its
descendants were ignored.

git-svn-id: http://trac-hacks.org/svn/tracjsganttplugin/0.11@11402 7322e99d-02ea-0310-aa39-e9a107903beb